Remove middle-ware for has_kicad_symbol_defined implementations
